add confirm alert request

+++ src/app/components/article-details/article-details.component.ts
@@ -2,6 +2,7 @@ import {Component, Input, OnInit} from '@angular/core';
 import {Article} from "../../model/Article";
 import {ActivatedRoute} from "@angular/router";
 import {ArticleService} from "../../services/article.service";
+import {formatCurrency} from "@angular/common";
 
 @Component({
   selector: 'app-article-details',
@@ -17,12 +18,13 @@ export class ArticleDetailsComponent implements OnInit {
 
   ngOnInit(): void {}
 
-  deleteArticle(id: number) {
-    this.articleService.deleteArticleById(id).subscribe(val => {
-      alert("Article was deleted");
-      window.location.assign('http://localhost:4200/show-articles')
+  deleteArticle(id: number, name: string) {
+    if (window.confirm('Are sure you want to remove ' + name.toUpperCase() + '?')) {
+      this.articleService.deleteArticleById(id).subscribe(val => {
+        alert("Article " + name.toUpperCase() + " was removed from your Cards!");
+        window.location.assign('http://localhost:4200/show-articles')
       });
-
+    }
   }
 
   openUpdateFormBool: boolean = false;
